Take into account Draft PRs
Draft PRs are used heavily by some tools, as an alternative to iterating in private.  Only Ready for Review PRs have been indicated by the user that they want to publish to GitHub for review.  This is a more normalized quantity across tools, and a bettter denominator for "Success Rate".  If data was available on all tasks created privately in each tool, then this could be compared with Draft PRs as well, but that data is not available publicly in GitHub.